Moving CommTensor from tests to private _spmd folder
[ghstack-poisoned]
