+++ torch/distributed/_spmd/__init__.py
@@ -1,0 +1,1 @@
+from .comm_tensor import _wait_comm, CommTensor
